Move intersection_search_space to avoid circular imports
