+++ examples/samplers/simulated_annealing_sampler.py
@@ -18,6 +18,7 @@ import numpy as np
 import optuna
 from optuna import distributions
 from optuna.samplers import BaseSampler
+from optuna.samplers.search_space import intersection_search_space
 from optuna import structs
 
 
@@ -31,7 +32,7 @@ class SimulatedAnnealingSampler(BaseSampler):
         self._current_trial = None
 
     def infer_relative_search_space(self, study, trial):
-        return optuna.samplers.intersection_search_space(study)
+        return intersection_search_space(study)
 
     def sample_relative(self, study, trial, search_space):
         if search_space == {}:
