+++ tests/integration_tests/test_skopt.py
@@ -5,6 +5,7 @@ from skopt.space import space
 
 import optuna
 from optuna import distributions
+from optuna.samplers.search_space import intersection_search_space
 from optuna.structs import FrozenTrial
 
 if optuna.type_checking.TYPE_CHECKING:
@@ -88,7 +89,7 @@ def test_is_compatible():
     study = optuna.create_study(sampler=sampler)
 
     study.optimize(lambda t: t.suggest_uniform('p0', 0, 10), n_trials=1)
-    search_space = optuna.samplers.intersection_search_space(study)
+    search_space = intersection_search_space(study)
     assert search_space == {'p0': distributions.UniformDistribution(low=0, high=10)}
 
     optimizer = optuna.integration.skopt._Optimizer(search_space, {})
