+++ tests/samplers_tests/test_samplers.py
@@ -226,47 +226,6 @@ def test_sample_relative():
         assert trial.params == {'a': 3.2, 'b': 'baz', 'c': 30, 'd': 30, 'e': 30}
 
 
-def test_intersection_search_space():
-    # type: () -> None
-
-    study = optuna.create_study()
-
-    # No trial.
-    assert optuna.samplers.intersection_search_space(study) == {}
-
-    # First trial.
-    study.optimize(lambda t: t.suggest_int('x', 0, 10) + t.suggest_uniform('y', -3, 3), n_trials=1)
-    assert optuna.samplers.intersection_search_space(study) == {
-        'x': IntUniformDistribution(low=0, high=10),
-        'y': UniformDistribution(low=-3, high=3)
-    }
-
-    # Second trial (only 'y' parameter is suggested in this trial).
-    study.optimize(lambda t: t.suggest_uniform('y', -3, 3), n_trials=1)
-    assert optuna.samplers.intersection_search_space(study) == {
-        'y': UniformDistribution(low=-3, high=3)
-    }
-
-    # Failed or pruned trials are not considered in the calculation of
-    # an intersection search space.
-    def objective(trial, exception):
-        # type: (optuna.trial.Trial, Exception) -> float
-
-        trial.suggest_uniform('z', 0, 1)
-        raise exception
-
-    study.optimize(lambda t: objective(t, RuntimeError()), n_trials=1, catch=(RuntimeError,))
-    study.optimize(lambda t: objective(t, optuna.exceptions.TrialPruned()), n_trials=1)
-    assert optuna.samplers.intersection_search_space(study) == {
-        'y': UniformDistribution(low=-3, high=3)
-    }
-
-    # If two parameters have the same name but different distributions,
-    # those are regarded as different trials.
-    study.optimize(lambda t: t.suggest_uniform('y', -1, 1), n_trials=1)
-    assert optuna.samplers.intersection_search_space(study) == {}
-
-
 @parametrize_sampler
 def test_nan_objective_value(sampler_class):
     # type: (typing.Callable[[], BaseSampler]) -> None
